Move PLUGIN_SELECTORS and its OSS defaults into whitelabel
The whitelabel module now owns the whole seam: the slot contract,
the OSS default implementations, and the late-binding reads.
EE plugs in through the module's barrel instead of metabase/plugins.

The composed reinitialize in metabase/plugins resets the slots through
a registered callback, since the registry cannot import upward to reach
the module's defaults.
