+++ enterprise/frontend/src/metabase-enterprise/settings/selectors.ts
@@ -1,7 +1,7 @@
 import noResultsSource from "assets/img/no_results.svg";
-import type { IllustrationValue } from "metabase/plugins";
 import type { State } from "metabase/redux/store";
 import { getSetting, getSettings } from "metabase/settings";
+import type { IllustrationValue } from "metabase/whitelabel";
 import type {
   EnterpriseSettings,
   IllustrationSettingValue,
